+++ src/Category/Changelog.md
@@ -10,6 +10,7 @@
 - Added the `taxonsIn()` and `firstTaxonIn()` methods to the `HasTaxons` trait.
 - Added the `is_active` field to the `Taxon` model/table
 - Added the `activeRootLevelTaxons()` method to the `Taxonomy` model
+- Added the `activeChildren` relationship to the `Taxon` model
 
 ## 5.x Series